+++ src/palm/runtimes/cli_pkg/bootstrap.py
@@ -12,7 +12,7 @@ import palm.patterns
 import palm.providers
 import palm.storages  # noqa: F401
 from palm.core.storage import StorageEngine
-from palm.executions.exceptions import DefinitionNotFoundError
+from palm.common.exceptions import DefinitionNotFoundError
 from palm.runtimes.cli_pkg.context import CliContext
 from palm.runtimes.embedded import EmbeddedRuntime